+++ test/test_schematic_solidmodel.jl
@@ -136,6 +136,19 @@ function test_component(component, clip_area, mesh=false, gui=false)
         SemanticMeta(:circle)
     )
 
+    # Add a NORENDER_META element that should be skipped in solid model rendering
+    render!(
+        floorplan.coordinate_system,
+        Rectangle(2μm, 2μm) + Point(50μm, 50μm),
+        DeviceLayout.NORENDER_META
+    )
+    # Also an element in a layer that will be designated as ignored by the target
+    render!(
+        floorplan.coordinate_system,
+        Rectangle(2μm, 2μm) + Point(50μm, 50μm),
+        SemanticMeta(:ignored)
+    )
+
     check!(floorplan)
     build!(floorplan)
 
@@ -160,6 +173,7 @@ function test_component(component, clip_area, mesh=false, gui=false)
         substrate_layers=[:chip_outline],
         levelwise_layers=[:chip_outline],
         indexed_layers=[:port],
+        ignored_layers=[:ignored],
         postrender_ops=[
             (
                 "substrates",
@@ -181,7 +195,9 @@ function test_component(component, clip_area, mesh=false, gui=false)
             ("substrates", 3),
             ("vacuum", 3),
             ("base_metal", 2),
-            ("circle", 2)
+            ("circle", 2),
+            ("norender", 2),
+            ("ignored", 2)
         ],
         solidmodel=true,
         simulation=true
@@ -201,6 +217,10 @@ function test_component(component, clip_area, mesh=false, gui=false)
     @test !SolidModels.hasgroup(sm, "writeable_area", 2)
     @test !SolidModels.hasgroup(sm, "base_negative", 2)
     @test !SolidModels.hasgroup(sm, "circle", 2)
+
+    # Verify that NORENDER_META element is not present (should be skipped even though it's retained)
+    @test !SolidModels.hasgroup(sm, "norender", 2)
+    @test !SolidModels.hasgroup(sm, "ignored", 2) # Same for `ignored_layers=[:ignored]`
 
     # The physical groups should be reindexed in decreasing order of dimension, and then
     # alphabetically